further folder restructure: as app is growing, a divide is needed between unique/general components
