+++ src/pageContent/spaceships/article/Article.module.css
@@ -2,7 +2,7 @@
   display: flex;
   flex-direction: column;
   flex: 1;
-  max-height: 440px;
+  max-height: 450px;
   max-width: 960px;
   margin: 1rem auto;
   text-align: center;
@@ -15,13 +15,13 @@
   );
   box-shadow: 2px 4px 4px rgba(46, 46, 46, 0.5);
   line-height: 1.5;
-  border-radius: 15px;
+  border-radius: 5px;
   border-bottom: 2rem solid transparent;
   box-sizing: border-box;
 }
 
 .content h3 {
-  border-radius: 15px 15px 0 0;
+  border-radius: 5px 5px 0 0;
   background: linear-gradient(
     to bottom right,
     rgba(255, 255, 255, 0.7),
