+++ src/pageContent/spaceships/details/ShipDetails.module.css
