+++ src/pageContent/spaceships/details/StaticInfoCard.module.css
