+++ src/pageContent/spaceships/list/SpaceShipsList.tsx
@@ -1,5 +1,5 @@
 import SpaceShip from '../../../models/SpaceShip';
-import List from '../../List';
+import List from '../../../components/list/List';
 import ShipItem from './ShipItem';
 import css from './SpaceShipsList.module.css';
 
@@ -18,14 +18,14 @@ export default function SpaceShipsList({ spaceships, className }: ListProps) {
   const liTransition = { type: 'tween', duration: 0.5 };
 
   return (
-      <List
-        className={css[className]}
-        items={spaceships}
-        keyFn={({ id }) => id}
-        liVariants={variants}
-        liTransition={liTransition}
-      >
-        {(item) => <ShipItem className={className} {...item} />}
-      </List>
+    <List
+      className={css[className]}
+      items={spaceships}
+      keyFn={({ id }) => id}
+      liVariants={variants}
+      liTransition={liTransition}
+    >
+      {(item) => <ShipItem className={className} {...item} />}
+    </List>
   );
 }
