+++ src/pageContent/spaceships/slideshow/SlideShow.module.css
