+++ src/pageContent/spaceships/slideshow/SlidesInfoCard.module.css
